% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioWearable-package.R
\docType{package}
\name{PhysioWearable-package}
\alias{PhysioWearable}
\alias{PhysioWearable-package}
\title{PhysioWearable: Free-Living Wearable Accelerometry for the Physio Ecosystem}
\description{
Free-living physical-activity analysis of raw tri-axial accelerometer data, following the GGIR methodology. Computes the Euclidean-Norm-Minus-One (ENMO) acceleration metric with epoch aggregation, gravity-based auto-calibration (van Hees et al. 2014), non-wear detection, step detection, physical-activity intensity classification and bout detection, and ingests raw accelerometer exports (GENEActiv / ActiGraph / Axivity CSV) into PhysioExperiment objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioWearable}
  \item \url{https://x-biosignal.r-universe.dev/PhysioWearable}
  \item \url{https://x-biosignal.github.io/PhysioWearable/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioWearable/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
